requirements.txt
streamlit
openai
pinecone
langchain
python-dotenv
